requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
deltalake==0.25.5
idna==3.10
numpy==2.2.6
pandas==2.2.3
pyarrow==20.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
six==1.17.0
tzdata==2025.2
urllib3==2.4.0